Write zone maps
Add functionality that will write each class of zones as a singleband
raster. The value of each shape in the raster is the zone ID. The
intention is that a user will open the maps in a GIS viewer, select
the zone of interest and then find the corresponding plot based on
the class and ID.

This required refactoring some more utility functions.