Custom values for initial res model
Can provide a custom res value per depth range by providing
the res_initial_custom_values and res_initial_custom_depth_ranges
arguments when constructing Model object.

Values is a list of resistivity values, and depth ranges is a list
of min, max pairs specifing a depth range. When make_mesh is called,
each value will be assigned to the corresponding depth range by list
position. Any depths not specified will be filled with the default
value (provided as 'res_initial_value' or 'res_model').

For this to work, you also need to set the Model init argument
res_initial_method to 'by_range'. The argument for original behaviour
of a single constant res is 'constant' and this is the default.