Custom depth layer growth
Can provide a series of depth ranges (as edges which will be made into
pairs inlcuding z1 and target_depth, e.g. [100, 500] corresponds
to ranges z1 to 100, 100 to 500 and 500 to target_depth) and
corresponding growth factors.

Growth factors are provided as list with length len(depth_ranges) + 1
(one factor for each depth range). Each factor is a tuple of
(value, type).

Type can be 'c' for constant, 'f' for factor or 'l' for log spacing.
Constant takes a value which is added to the cell to grow the depth.
Factor takes a value which is multiplied by the cell depth to get the
next cell. Log spacing takes the number of of log-spaced layers to
generate in that depth range.

This is only one possible approach. I'm waiting on feedback to see if
this is acceptable. An issue is how to control the growth of cells not
specified in a depth range. This approach eliminates that by making
the user provide range boundaries, so the whole depth range is always
covered.