Zones by polygons + fix gdal origin bug
Can now provide a GeoSeries of polygons (with name of area as index)
to find_zones and this will divide a model into zones based on cells
intersecting polygons.

Also fixed a bug with get_gdal_origin. Wasn't accounting for
+south CRSs. This is fixed and have applied the fix to
convert_modem_to_geogrid.

Next is to add the ability to load the polygons from shapefiles.